move the helpers and tests so that we don't have to import the script itself
