+++ packages/monorepo-tools/src/index.ts
@@ -8,3 +8,4 @@ export * from './utils/with-progress';
 export * from './utils/workspace-dependencies';
 export * from './utils/get-packages-in-topological-order';
 export * from './utils/get-npm-token-list';
+export * from './utils/package-helpers';
